Intial move of files
